Move minime token tests to their own folder
